pyproject.toml
[tool.poetry]
name = "xk-backend"
version = "0.1.0"
description = ""
authors = ["liupengfei <francis.liu@rykjsz.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "3.2"
djangorestframework-simplejwt = "^5.0.0"
pillow = "^8.4.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-cors-headers = "^3.10.1"
dynaconf = "^3.1.7"
django-debug-toolbar = "^3.2.4"
psycopg2 = "^2.9.2"
daphne = "^3.0.2"
gunicorn = "^20.1.0"
pandas = "^1.3.5"
django-extensions = "^3.1.5"
django-lifecycle = "^0.9.3" # 添加像Signal一样的生命周期钩子
shortuuid = "^1.0.8"
uvicorn = "^0.17.5"
nplusone = "^1.0.0"
drf-nested-routers = "^0.93.4"
boto3 = "^1.21.22"
django-storages = "^1.12.3"
drf-extensions = "^0.7.1"
django-safedelete = "^1.1.2"
openpyxl = "^3.0.9"
django-denorm-iplweb = "^0.5.4"
django-mptt = "^0.13.4"
django-redis = "^5.2.0"
faker = "^13.12.0"
model-bakery = "^1.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-django = "^4.5.2"
model-bakery = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = 'aliyun.mirrors'
url = "http://mirrors.aliyun.com/pypi/simple/"
